requirements.txt
scrapy>=2.11
python-dotenv>=1.0
parsel>=1.9
lxml>=5.3
